switched to HSB model of line highlighting
Now, the saturation gradually falls off over time. Also, the brightness goes up over time, but it starts much higher. As a result, changes introduced by the current commit tend to be a somewhat dark but bold yellow. Changes in slightly older commits are a lighter but much less bold yellow. As time goes on, the changes eventually blend to white. There is no maxage anymore; the function to compute both saturation and brightness is a variation of 1/age, though this means that there is an effective maximum age beyond which your monitor cannot display a difference.